move getport to commons file - u9g
